Cond now takes 1 argument, also fix RetrieveAll bug
At one point I may have been thinking about a Between version of the
Cond command, but since I now have the Between call and Cond not having
an applicable Between operator, it is useless for me to have Cond take a
variable number of parameters. I removed it so the single parameter Cond
calls would work. Also there was a bug in RetrieveAll where if you had
existing elements in the passed array, you would get the new elements
appended to the existing elements, it should replace the array and the
only elements to be found would be the ones you searched for.