Add the mixin functionality, add simple Count function
Mixins allow some degree of ActiveRecord (pattern not the Rails thing)
functionality to be provided by the db library. At the moment, you can
Save, Delete, and UpdateAttribute(s) on initialized struct instances.
Since we need a pointer to at least the item, we can't just magically
initialize along with the struct we're mixed in, but I think the current
method of getting initialized is the best possible compromise at the
moment.