Document Configs, remove unused code from new_mapper
Document the ColumnInfo struct that is used in Dialects, also add
documentation around the different Config fields. IdName made send to
change to a more expressive function from a static field. Also found out
that one of the users of IdName was left behind in a refactoring and
never removed, so that code's remove now.