Make FieldToColumn more expressive
I've heard that a pattern (not a good pattern, but a pattern) of naming
database columns it to name it TableName_ColumnName to make Joins and
aggregations easier to read when only listing column names. In order to
support patterns like that, we'll pass the struct name and field name to
the FieldToColumn function.