Add value Id extraction
Let's say you want to find all posts that were written by an array of
User's, but don't want to create a new array and a for loop just to make
that happen. In that case, just pass in the array of User structs, db
can now extract the ids from arguments to queries so long as those
structs have registered mappers for them.